174
몬스터 상태 업데이트 오류들 수정 및 코드 개선

+++ Engine_SOURCE/Player.cpp
@@ -156,7 +156,7 @@ namespace dru
 		}
 		
 
-		for (size_t i = 0; i < mAfterImages.size(); i++)
+		for (UINT i = 0; i < mAfterImages.size(); i++)
 		{
 			mAfterImages[i]->SetIndex(i);
 		}
